+++ CHANGELOG.md
@@ -9,7 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- convertStreamToString
+- `convertStreamToString` function
+
+## 2.5.0 - 2023-05-17
+
+### Added
+
+- CIP68 RFT 444 support
+
+## 2.4.0 - 2023-03-15
+
+### Removed
+
+- `addr_vk` support in `paymentCredToBech32Address`
 
 ## 2.3.2 - 2023-03-07